Implementing literals
If no entry for a given key is found in the Dictionary, we consider the key itself as a string literal that should be inserted at the given point. E.g:

phrase 'The', 'noun', 'verb', 'preposition', 'a', 'noun', '.'

If no rules are found for 'The', it is appended to the output.