Defining methods instead of using Dictionary
Instead of maintaining a Dictionary with entries for each word class, we can (dynamically) define methods for them instead. Since the responsibility of transforming a key (e.g. 'noun') into a word (e.g. 'dog') is now in the hands of the dynamically defined methods, we no longer need to fetch & append each key when applying a rule, so Rule.apply now simply joins strings together. Methods wit a bang (!) add a period after the chosen word.

This is probably not a good idea for any real application which depends on a database, but is pretty damn cool.